fix: export createAnonymousContext define
