+++ src/lib/application.ts
@@ -234,22 +234,6 @@ export class Application extends EggApplicationCore {
   }
 
   /**
-   * Run async function in the anonymous context scope
-   * @see Context#runInAnonymousContextScope
-   * @param {Function} scope - the first args is an anonymous ctx, scope should be async function
-   * @param {Request} [req] - if you want to mock request like querystring, you can pass an object to this function.
-   */
-  async runInAnonymousContextScope(scope: (ctx: Context) => Promise<void>, req?: unknown) {
-    const ctx = this.createAnonymousContext(req);
-    if (!scope.name) {
-      Reflect.set(scope, '_name', eggUtils.getCalleeFromStack(true));
-    }
-    return await this.ctxStorage.run(ctx, async () => {
-      return await scope(ctx);
-    });
-  }
-
-  /**
    * secret key for Application
    * @member {String} Application#keys
    */
